+++ inst/doc/shinyFeedback-intro.R
@@ -26,3 +26,31 @@
 #  
 #  shinyApp(ui, server)
 
+## ---- eval = FALSE-------------------------------------------------------
+#  ui <- fluidPage(
+#    useShinyFeedback(), # include shinyFeedback
+#  
+#    numericInput(
+#      "multiFeedbacks",
+#      "1 is scary 2 is dangerous",
+#      value = 1
+#    )
+#  )
+#  
+#  server <- function(input, output) {
+#    observeEvent(input$multiFeedbacks, {
+#      feedbackWarning(
+#        inputId = "multiFeedbacks",
+#        condition = input$multiFeedbacks >= 1,
+#        text = "Warning 1 is a lonely number"
+#      )
+#      feedbackDanger(
+#        inputId = "multiFeedbacks",
+#        condition = input$multiFeedbacks >= 2,
+#        text = "2+ is danger"
+#      )
+#    })
+#  }
+#  
+#  shinyApp(ui, server)
+