+++ NEWS.md
@@ -1,3 +1,10 @@
+# shinyFeedback 0.0.2
+
+- support for `shiny::selectInput()`
+- feedback messages are not longer displayed for non supported `shiny::*Input()`s
+- support for multiple feedbacks for a single shiny input
+- Font Awesome icons can now be used
+
 # shinyFeedback 0.0.1
 
 - Initial release of shinyFeedback